refactor(checker): split diagnostic source span queries
